# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Imóveis",
        "description": "API para gerenciamento de imóveis e contatos",
        "contact": {
            "name": "Suporte",
            "email": "suporte@exemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
        "/contacts": {
            "post": {
                "tags": [
                    "Contato"
                ],
                "summary": "Enviar mensagem de contato",
                "description": "Endpoint para enviar uma mensagem de contato através do sistema",
                "operationId": "sendContact",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "message"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 300,
                                        "example": "joao.silva@exemplo.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "(11) 99999-9999"
                                    },
                                    "message": {
                                        "type": "string",
                                        "maxLength": 300,
                                        "example": "Gostaria de mais informações sobre o imóvel"
                                    },
                                    "terms_accept": {
                                        "description": "Confirmação de aceite dos termos de uso e política de privacidade",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "recaptchaToken": {
                                        "description": "Token do Google reCAPTCHA v3 para validação de segurança",
                                        "type": "string",
                                        "example": "03AFcWeA5X..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Mensagem enviada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "contacts"
                                                    ]
                                                },
                                                "attributes": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "João Silva"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "joao.silva@exemplo.com"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "(11) 99999-9999"
                                                        },
                                                        "message": {
                                                            "type": "string",
                                                            "example": "Gostaria de mais informações sobre o imóvel"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Em breve retornaremos seu contato."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na validação do reCAPTCHA",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Falha ao verificar o recaptcha."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os dados fornecidos são inválidos."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Como podemos te chamar"
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Precisamos saber seu e-mail, para que possamos entrar em contato."
                                                    }
                                                },
                                                "message": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Descreva em poucas palavras: sua dúvida, mensagem ou sugestão."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}